% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authordown_validate.R
\name{authordown_canonicalize}
\alias{authordown_canonicalize}
\title{Standardize raw author metadata}
\usage{
authordown_canonicalize(data)
}
\arguments{
\item{data}{A data frame of author metadata.}
}
\value{
A standardized data frame.
}
\description{
Standardize raw author metadata
}
\keyword{internal}
